Move: `IAnalysisRepository.cs` in `Interfaces` folder
