+++ Clinical.UseCases/UseCases/Analysis/Commands/CreateCommand/CreateAnalysisHandler.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using Clinical.Interface;
+using Clinical.Interface.Interfaces;
 using Clinical.UseCases.Commons.Bases;
 using MediatR;
 using Entity = Clinical.Domain.Entities;
