+++ Clinical.UseCases/UseCases/Analysis/Queries/GetAllQuery/GetAllAnalysisHandler.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
 using Clinical.Application.DTOS.Analysis.Response;
-using Clinical.Interface;
+using Clinical.Interface.Interfaces;
 using Clinical.UseCases.Commons.Bases;
 using MediatR;
